initial checkin.
This is a simple app that provides a configuration screen for setting
the UHD properties of a GNU Radio app over control port. Only supports
the 'freq' command right now, but eventually will have a drop-down for
all UHD commands. The value for the command will also be augmented to
have a spinner box that allows you to set the unit (e.g., Hz, kHz,
MHz, etc. for the freq command).